Updated window-size detection hook (useDetectWindowSize) to conform to react standards.
